requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
blinker==1.8.2
cffi==1.16.0
click==8.1.7
Flask==3.0.3
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-PageDown==0.4.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-Toastr==0.5.8
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycparser==2.22
SQLAlchemy==2.0.30
typing_extensions==4.12.1
Werkzeug==3.0.3
WTForms==3.1.2
